Fix #114855: Crash Loading Win32 Quick Access
Replace use of macro FAILED with explicit checks for S_OK returns. Some
of these functions can return S_FALSE, which is a negative condition
that is not failure. It is not caught by the FAILED macro. Therefore
this function will fail in circumstances where the quick access folder
is not found or perhaps empty.