Cleanup: Synchronizing Icons
Mostly just the result of running `blender_icons_update.py` without adding or changing any icon source files. Removal of three unused icons and adding three others to the project to resolve conflicts between them.

---

The three icons removed (icon_lightprobe_cubemap, icon_lightprobe_planar, icon_lightprobe_grid) are in CMakeLists.txt but are not defined as icons or used. The three icons added to the project (icon_lightprobe_plane, icon_lightprobe_sphere, icon_lightprobe_volume.dat) are created in the file system by blender_icons_update.py but not added to the project. These icons are used.

Best guess is that icon_lightprobe_plane, for example, is meant to replace icon_lightprobe_planar but the old one was not removed and the new one not added to the project.