Add Glsl language for text editor
Pr adds the Glsl language to the text editor.
The objective is to be able to create your own GLSL shaders in blender. This improves the workflow since all shader programming can be done directly in Blender. In addition, the GLSL language is a very popular language in the video games industry and even in general.
The goal is then to also support HSLS in the same way in order to have a complete solution with OSL, GLSL and HLSL