Retopology: Use Material Colors In Overlay
This PR adds an option to the retopology overlay to switch between the theme
color or the material color. It is a common workflow to color faces based
on their importance during retopology. This change allows you to see the
colors during the retopology process.

Request from the Blender Studio. Implements design #114843.