Vulkan: Make Command Pool Context Specific
In Blender a context should not be shared between threads. In Vulkan a
command pool must not be shared between threads. In the current
implementation the command pool was stored on device level and was
shared between multiple context which made the implementation not
multi-thread aware.

This PR moves the command pool from device to command buffers where it
would not conflict between other contexts.