Vulkan:Refactor:Renderpass Configuration for Eevee
Setting the `GPU_framebuffer_texture_attach` function is often ambiguous. In that case, use `framebuffer_config`.
Also, unfortunately, if there is a difference between Image-Format and Image-View-Format, we will need to regenerate the render pass.
It is necessary to clarify what to improve efficiency with `GPU_framebuffer_texture_attach`.
we must regenerate the render pass each time we change the image format or number of attachments.
Most of the Format changes are to sRGB, but this is still something that needs to be verified.