BLI_fileops: Harmonize 'rename' behaviors accross platforms.
This commit tries to make the behaviors of `BLI_rename` and
`BLI_rename_overwrite` more consistent and coherent accross all
supported platforms.

* `BLI_rename` now only succeeds in case the target `to` path does not
  exists.
* `BLI_rename_overwrite` allows to replace an existing target `to` file
  or (empty) directory.

NOTE: In case the target is open by some process on the system, trying
to overwrite it will still fail on Windows, while it should succeed on
Unix-like systems.

The main actual change is for Windows, with the usage of `MoveFileExW`
instead of `_wrename`, which allows for 'native support' of file
overwrite (using the `MOVEFILE_REPLACE_EXISTING` flag). Directories
still need to be explicitely removed though.