Fix: DEG: Image updates are flushed for initial evaluation
Visibility updates as part of the first depsgraph evaluation when
data-blocks first become visible are flushed to editors for Image IDs.

A check does exist to avoid flushing in those cases, however, it uses
the COW expanded status of the ID, which fails for IDs that do not use
COW.

This patch fixes that by explicitly passing the original value of the
graph visibility update flag before getting reset, which is then used in
the aforementioned check.