Python: Expose `NodeLink.multi_input_socket_index`
This exposes `NodeLink.multi_input_socket_index` as a readonly property,
along with a `swap_multi_input_socket_index` function that provides a
safe and composable way to change it from Python.

Additionally this property is used in `NodeSocket.links` to sort the
returned tuple for input sockets.
This fixes a bug in Node Wrangler, where resetting nodes could change
the order of links.

This allows addons to read and reorder inputs to multi socket nodes,
functionality that was previously missing and could only be done from
C++, despite not being an uncommon operation.