Fix: bevel weights in Simple mode of Solidify
In Simple Solidify's mode  edge weights created by "Bevel Convex" disappear, if edge weight is added to original mesh. From technical side problem is identical to the one in #114860.
Second problem was that vertex bevel weights are ignored by Solidify Simple. Code in patch transfers bweight of the vertex to the edge bweight extruded from it.