VSE: optimize draw_channels by speeding up rna_SeqTimelineChannel_owner_get
Drawing the channels list was taking about 3ms on my machine, almost all
of that time querying *all* the VSE strips for each channel mute/lock
button. That is done due to some complexicated code from, where deep
from inside button it calls ui_but_get_fcurve which eventually calls
rna_SeqTimelineChannel_path, which calls rna_SeqTimelineChannel_owner_get,
which was proceeding to construct a set of all the strips.

It's a good question whether this "all the strips" set could be cached
somewhere.

But for now, just change the code to not query all the strips, but
instead query a set of all meta strips, since that's what the code is
only interested in anyway. This makes draw_channels go from 3.0ms
down to 1.3ms.