UI: dont run update functions if button values did not change
Symptoms were recalculation in cases were these are not needed (e.g.
putting the same value as before on a heavy modifier property would give
a delay, same as selecting e.g. the same vertexgroup that was previously
selected). As shown in #40009, this delay could also end up swallowing
doublick-events, so on a heavy mesh e.g. a vertexgroup could not be
renamed by doublicking on it.

The doubleclick issue can be solved with event timestamps (see
efef709ec73c for this being done for linux), however the - unneccessary
- delay would still take place.

This PR skip RNA property set/update functions in case a button value
did not change, thus eliminating the delay.

It is basically the second try at this (see eb06ccc32462 which got
reverted), but this time leaving the button handling code untouched.

Fixes #87448
Fixes #40009