Vulkan:Fix:Resource Discard
The process of deferring the discard of a resource should be done on a per-context basis.
This is because if multiple contexts are running and one context destroys all resources with end_frame(),
there will be a context that cannot access that resource.

As a prerequisite, when a deconstructor is called for a resource,
the system must be built so that no context other than the one
that did the destruction work already uses that resource.
This should have already been done with the responsibility of the draw module.