Realtime Compositor: Add static cached images
The realtime compositor currently relies on the GPU cache in image IDs.
That cache only supports single layer images, so multi-layer images will
be acquired without a cache, introducing significant IO bottlenecks for
the GPU compositor. Furthermore, the image ID GPU cache stores images in
formats that are not compatible with the realtime compositor, producing
erroneous alpha handling and data loss.

This patch ignores the image GPU cache and stores the images in the
static cache manager of the compositor. Draw data was introduced to the
image ID for proper cache invalidation, like other IDs such as masks.
The cached images are ensured to be premultiplied without data loss by
acquiring image buffers with packed alpha, then handling alpha manually.