Fix: allow overshoot numerical input in the Pose Breakdowner operator
You could already do overshoot via the mouse, but it arbitrarily
disallowed when entering values numerically.  This simply lifts that
restriction so that numerical input is capable of the same things as
dragging the mouse.

Importantly, the operator slider code already appropriately clamps
both mouse and numerical input when needed.  So the checks and clamping
removed in this commit simply aren't needed, and removing them just
makes numerical input get constrained exactly the same as mouse input.