GPU:Add:Subpass API
Adds a function that provides the information needed to implement multiple subpasses in Vulkan as bit flags.
Provides the texture with the information necessary to build it.
This will allow it to be recognized when configuring the framebuffer.
Its contents are write paths and slots. Also, if there is a read, the path and slot of the read.
Vulkan has a Renderpass image description order and a subpass reference order. This plus Blender's AttachmentType order.
The image description order is the order fixed during the `framebuffer_config_array` function.
What is required for setting is the `subpass number` and `subpass reference order`(=shader location number).