Vulkan:Add:VKSubpassDescription
Determine Subpass dependencies from the enumerated type.
InputAttachment is a strict case, so there is little variation in access patterns.
However, for the purpose of this rationalization, we will build it so that it always transitions to the best layout.
This is a low-overhead process because it is embedded in the pass. (compared to barrier)