Vulkan:Fix:Frambuffer test
I've tweaked it so that the test passes.
Here's an unclear point. The fact that Flush is submit and finish is wait may not be guaranteed yet.
Next, let's talk about the usage of Multiview.
Programs that use gl_viewIndex can be embedded in Subpass.
It is typically stereo rendering. It is also suitable for CubeMap drawing.
This test is a multi-viewports that pertains to pipelines.
Is gl_viewindex ever used in connection with subpass?