USD: Add option to export only deform bones for armatures
This patch adds an Only Deform Bones flag to the exporter so that
when exporting UsdSkel assets, non-deforming helper bones are
skipped.  If deform bones live as leaf bones along a  chain of
non-deforming bones, their non-deforming parents are also exported
in order to preserve hierarchy relationships and animation.

Finally, this patch reorders the joint weight binding orders for
UsdGeomMesh exports to match the order and count of joints
written on the UsdSkel side per clip.

This commit was primarily authored by Charles Wardlaw with edits
by Michael Kowalski.