Fix #115662: Outliner slows down many geometry operations in big scenes
This was reported for UV editing but also e.g. some modeling operations
were affected.

In 295bc1249a85, a listener for `NC_GEOM` > `ND_DATA` was added to the
Outliner in order to update when renaming bones (c4622c405e40).
Since we are only interested in the naming part (no need to update the
Outliner otherwise), the notifier/listener combo was made more specific
by including the `NA_RENAME` action here.

Since this was a very general thing to listen to, other operations might
have relied on this to properly update, but having checked many things,
I could spot only one case where an Outliner update was missing after
the initial change and that was adding images (this was notifiying XXX).
This case was added separately now.