Vulkan :Refactor :Layout Policy
Establish a policy on the layout of VkImage.
When the layout is not being processed,
it should always be this layout.
The good thing about this is that you don't have to change the layout during the binding of the resource.
Plus, there's no need to keep track of layouts. This PR is related to the streamlining of Renderpass.
The dummy attachment should be removed in a later PR.