GPU :Refactor :Framebuffer config
Make the function that configures the frame buffer from the GPU module a virtual function.
Need for this change:
Vulkan's render passes are a concept within the framebuffer concept assumed by the GPU module.
This is evidenced by the fact that running the Framebuffer_config_array function provides all the information needed to generate the Vulkan-Render-pass.
Therefore, we believe that it is necessary to utilize virtual functions to configure functions suitable for each backend.
This PR does not change any substantive content.