BMesh: Optimize copying attributes from many elements at once
Fixes #115776

Create a minimal structure that contains the instructions necessary to
copy from one custom data format to another. This structure is similar
to the one used in dfacaf4f409287baa1c8. It should have been used in
9175d9b7c27b191ea94a, which instead introduced quadratic performance
relative to the number of layers when copying every element.

A few areas in BMesh are changed to use this map:
- Copying the entire mesh
- Adding a custom data layer (attribute)

The non-map attribute copy functions are also changed to check for when
the source and result BMeshes are the same. In that case it's much
faster to call the "same format" function from 9175d9b7c27b191ea94a.

For numbers, the timings are arbitrarily influenced by how many layers
I add in my testing. With many very many layers, a 10x difference is
easily observable though.