Fix #100784: Truncate Text with Ellipsis inside Number Inputs
When a property name is shown inside of a numerical input, truncate
with ellipsis if it cannot fit.

---

We current don't truncate with ellipsis for `UI_BTYPE_NUM` and `UI_BTYPE_NUM_SLIDER`, but just remove characters. Probably because these have the complication of looking like "this is long text :0.00" so we have to deal with the trailing value.

This PR just adds an early exit if the string fits. If not the string measurement is done with the length of ellipsis considered. And then inserts that character at the end if it can.