Fix #109024: Off-by-1 in rna_access for non-array props without raw access
The `a + itemlen > in.len` check was off-by-1 whenever accessing a
non-array property without raw access. This was because `itemlen` was
actually the array length of the property, which is `0` for non-array
properties.

Given an array which was too short, this would cause the slower loop to
overrun the end of the array by one item, causing a crash on a debug
build with `Fatal Python error: _PyMem_DebugRawFree: bad trailing pad
byte`.

The semantics of the `itemlen` and `arraylen` variables have now been
swapped to be less confusing. So `arraylen` is now the array length of
the property and `itemlen` is now the item length of the property. The
`arraylen` variable has been given the same scope as `itemlen` so that
it can be used in the slower loop when raw access is not available.