RNA: Speed up raw array access when the property is stored contiguously
Recent and ongoing efforts have changed many properties to be stored
contiguously in memory, e.g. mesh attributes. This patch updates
rna_raw_access to make use of this and copy the entire contiguous block
of memory when the property is stored contiguously.

This is faster and scales much better with larger arrays.

Using foreach_get/foreach_set with the `value` property of a `'FACE'`
domain `'INT'` attribute:

Using a mesh with 98304 faces:

foreach_get with a compatible buffer object:
- Before: ~230μs
- After: ~8μs

foreach_get with a Python list:
- Before: ~1760μs
- After: ~1530μs

foreach_set with a compatible buffer object:
- Before: ~200μs
- After: ~8μs

foreach_set with a Python list:
- Before: ~970μs
- After: ~770μs

Using a mesh with 384 faces:

foreach_get with a compatible buffer object:
- Before: ~1.49μs
- After: ~0.55μs

foreach_get with a Python list:
- Before: ~7.35μs
- After: ~6.40μs

foreach_set with a compatible buffer object:
- Before: ~1.44μs
- After: ~0.57μs

foreach_set with a Python list:
- Before: ~4.43μs
- After: ~3.57μs