GPU: Enable Shader Compilation on During Buildbot Testing
This PR enabled running shader validation tests on buildbot during
tests. This is enabled for Vulkan and Metal. OpenGL isn't enabled
as OpenGL requires an actual driver and GPU attached to the build
bot infrastructure.

Metal can run as a Mac always has a GPU. For Vulkan we expect that we
can activate a software vulkan device. Compilation is responsibility
of Blender and doesn't rely on any driver specific configuration.

Per backend the shaders are compiled twice. Once with based on the
actual requirements of the installed GPU/backend. And once with all
the work-arounds enabled, simulating a platform close to the
minimum requirements of Blender.