Remove workaround for ARM64/WSL Mesa drivers
This reverts #110582 - the workaround is no longer needed, as MS have
updated the version of mesa they ship as part of their compatibility
pack, which contains a workaround on that side instead.

This means that the supported version of OpenGL based on the host GPU is
now correct, bar `GL_ARB_texture_view` on ARM64 devices.

`GL_ARB_texture_view` is faked for ARM64 devices, as the part that is
not implemented (channel casting) is not used in blender.

See: https://gitlab.freedesktop.org/mesa/mesa/-/commit/698344b93c49a9f3a257a0ef4546edf5cd3a9130