Fix #116075: OpenGL Shader Compilation On Selected Platforms
Due to recent changes an cached string in the GLShader grew out of
its bounds. This resulted in incorrect shader generation on selected
platforms (Reported was Windows/NVIDIA).

This PR replaces the old C-style string generation with CPP-style
string stream, making sure that the allocated memory grows with the
size of the string.