Transform: Test occlusion in individual Face Project if Nearest is also enabled
When Individual Face Project is enabled, occluded elements being
transformed are moved in front of the occluding object.

This is old behavior, and may even be desirable in some cases.

But when we enable Face Nearest together, it seems to be more
advantageous to use Nearest for occluded elements.

Therefore, this commit changes the behavior of Face Project + Nearest
so that Nearest is used not only for elements that don't hit an object,
but also for elements that are occluded.