Lookdev: Use Mid grey for Diffuse HDRI Sphere
When the lookdev balls where initially added we didn't had the
experient community than we currently have and reused the value we
used for our default materials.

Nowadays we have actually industry people asking to use a mid grey
albedo of 18% as that is widely used in the vfx studios. This PR
changes the default diffuse HDRI sphere to be 50% grey.

For references see:
- https://en.wikipedia.org/wiki/Middle_gray